+++ tests/table_header_rowspan.rs
@@ -7,7 +7,7 @@
 fn native_header_cell_spans_into_body() {
     assert_eq!(
         carve::to_html("|= H |= G |\n| ^ | b |\n| ^ | c |"),
-        "<table>\n  <thead><tr><th scope=\"col\" rowspan=\"3\">H</th><th scope=\"col\">G</th></tr></thead>\n  <tbody>\n    <tr><td>b</td></tr>\n    <tr><td>c</td></tr>\n  </tbody>\n</table>"
+        "<table>\n  <thead>\n    <tr><th scope=\"col\" rowspan=\"3\">H</th><th scope=\"col\">G</th></tr>\n  </thead>\n  <tbody>\n    <tr><td>b</td></tr>\n    <tr><td>c</td></tr>\n  </tbody>\n</table>"
     );
 }
 
@@ -15,7 +15,7 @@ fn native_header_cell_spans_into_body() {
 fn gfm_separator_header_cell_spans_into_body() {
     assert_eq!(
         carve::to_html("| H | G |\n|---|---|\n| ^ | c |"),
-        "<table>\n  <thead><tr><th scope=\"col\" rowspan=\"2\">H</th><th scope=\"col\">G</th></tr></thead>\n  <tbody>\n    <tr><td>c</td></tr>\n  </tbody>\n</table>"
+        "<table>\n  <thead>\n    <tr><th scope=\"col\" rowspan=\"2\">H</th><th scope=\"col\">G</th></tr>\n  </thead>\n  <tbody>\n    <tr><td>c</td></tr>\n  </tbody>\n</table>"
     );
 }
 
@@ -23,6 +23,6 @@ fn gfm_separator_header_cell_spans_into_body() {
 fn header_rowspan_and_body_rowspan_coexist() {
     assert_eq!(
         carve::to_html("|= H |= G |\n| ^ | b |\n| x | ^ |"),
-        "<table>\n  <thead><tr><th scope=\"col\" rowspan=\"2\">H</th><th scope=\"col\">G</th></tr></thead>\n  <tbody>\n    <tr><td rowspan=\"2\">b</td></tr>\n    <tr><td>x</td></tr>\n  </tbody>\n</table>"
+        "<table>\n  <thead>\n    <tr><th scope=\"col\" rowspan=\"2\">H</th><th scope=\"col\">G</th></tr>\n  </thead>\n  <tbody>\n    <tr><td rowspan=\"2\">b</td></tr>\n    <tr><td>x</td></tr>\n  </tbody>\n</table>"
     );
 }